addd cashier window added with branch details inout
